move DummyAction to resolver package
