+++ gshell-core/src/test/java/com/planet57/gshell/command/resolver/DummyAction.java
@@ -13,7 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.planet57.gshell.command;
+package com.planet57.gshell.command.resolver;
+
+import com.planet57.gshell.command.CommandAction;
+import com.planet57.gshell.command.CommandActionSupport;
+import com.planet57.gshell.command.CommandContext;
 
 import javax.annotation.Nonnull;
