+++ gshell-core/src/test/java/com/planet57/gshell/command/resolver/NodePathCompleterTest.java
@@ -19,7 +19,6 @@ import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Module;
 import com.google.inject.Stage;
-import com.planet57.gshell.command.DummyAction;
 import com.planet57.gshell.command.registry.CommandRegistry;
 import com.planet57.gshell.command.registry.CommandRegistryImpl;
 import com.planet57.gshell.event.EventManager;
