+++ gshell-core/src/test/java/com/planet57/gshell/command/resolver/NodeTest.java
@@ -19,7 +19,6 @@ import java.util.Collection;
 import java.util.Iterator;
 
 import com.planet57.gshell.command.CommandAction;
-import com.planet57.gshell.command.DummyAction;
 import com.planet57.gshell.command.GroupAction;
 import junit.framework.Assert;
 import org.junit.After;
